+++ src/test/java/com/se/fileserver/v1/application/service/FileDownloadServiceTest.java
@@ -8,7 +8,10 @@ import com.se.fileserver.v1.common.infra.security.provider.JwtTokenResolver;
 import com.se.fileserver.v1.file.adapter.controller.FileController;
 import com.se.fileserver.v1.file.adapter.presenter.FilePresenter;
 import com.se.fileserver.v1.file.application.dto.FileDownloadDto;
+import com.se.fileserver.v1.file.application.service.FileDeleteService;
 import com.se.fileserver.v1.file.application.service.FileDownloadService;
+import com.se.fileserver.v1.file.presentation.controller.FileController;
+import com.se.fileserver.v1.file.presentation.presenter.FilePresenter;
 import java.net.MalformedURLException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -46,6 +49,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 public class FileDownloadServiceTest {
 
   private MockMvc mockMvc;
+
+  @MockBean
+  private FileDeleteService fileDeleteService;
 
   @MockBean
   private FileDownloadService fileDownloadService;
